requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
mysqlclient==2.1.1
packaging==23.0
pluggy==1.0.0
protobuf==3.20.3
pytest==7.2.2
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.15
Werkzeug==2.2.3
